[MOB-865] Added missing progress bar text. Fixed progress bar drawable

+++ app/src/main/java/com/asfoundation/wallet/promotions/Promotion.kt
@@ -33,8 +33,8 @@ class ProgressItem(
     val title: String,
     val icon: String?,
     val endDate: Long,
-    val current: Double,
-    val objective: Double
+    val current: BigDecimal,
+    val objective: BigDecimal
 ) : Promotion(id)
 
 class GamificationItem(
